pyproject.toml
[tool.poetry]
name = "7x1"
version = "0.1.0"
description = ""
authors = ["levan <levan.bujiashvili.1@btu.edu.ge>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
boto3 = "^1.26.135"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
